Updated GameManagerSim to run a game from it's current point *without* it creating a new simulation from each potential play beyond it.
Basically if it's called when there are 2 cards on the board, it will
finish the rest of that game without creating a new simulation (during
that current simulation) when a third card is added to the board.

Also updated a bit of the output messages for debugging to include extra
info.